Workaround for pymssql import error
We try to re-install 'pymssql' pip package directly on the deployment
machine, if the bundled package with charm couldn't be imported.

The 'pymssql' package is installed with a cpython library, which is
pre-compiled for every supported Python version.

So, when we build the charm via 'charmcraft build', the built charm
will contain the cpython library corresponding to the Python version
used to build the charm.

The Python version from the deployment machine might not be the same,
and the charm will fail to import 'pymssql'.
For example: we build the charm on Ubuntu Focal with Python 3.8, and we
deploy it on Ubuntu Bionic with Python 3.6.

As a workaround, we try to re-install the charm 'requirements.txt'
dependencies on the deployment machine to fix this issue.